Add fallback runScript insertion location
There's an issue wherein having no script tags on the page (but having the
js in memory executing) causes errors since runScript has no reference
point to inject scripts.

This commit allows for two cases:

1) If there are no script tags on the page, it will grab the last element
node in the body and insert scripts next to that element.

2) If there are no script tags on the page and the body element is EMPTY,
a script tag will be created and appended to body. Subsequent calls to
runScript will allow the first use case of inserting next to scripts to be
called.